+++ src/pages/ModelizerDrawPage.vue
@@ -1,73 +1,43 @@
 <template>
-  <q-layout
-    container
-    class="modelizer-draw-view"
-    data-cy="modelizer-draw-view"
-  >
-    <component-definitions-drawer
-      v-if="data.plugin"
-      :plugin="data.plugin"
-      :templates="templates"
-      :project-name="projectName"
-    />
-    <q-page-container>
-      <q-page class="bg-grey-3">
-        <div
-          id="root"
-          data-cy="draw-container"
-          @dragover.prevent
-          @drop.prevent="dropHandler"
-        >
-          <component-drop-overlay />
-        </div>
-      </q-page>
-    </q-page-container>
-    <component-detail-panel
-      v-if="data.plugin"
-      :plugin="data.plugin"
-    />
-  </q-layout>
+  <q-page class="bg-grey-3">
+    <div
+      id="root"
+      data-cy="draw-container"
+      @dragover.prevent
+      @drop.prevent="dropHandler"
+    >
+      <component-drop-overlay />
+    </div>
+  </q-page>
 </template>
 
 <script setup>
+import ComponentDropOverlay from 'components/drawer/ComponentDropOverlay';
 import {
+  addNewComponent,
+  addNewTemplateComponent,
+  getPluginByName,
+  renderConfiguration,
+  renderModel,
+} from 'src/composables/PluginManager';
+import { ComponentDrawOption } from 'leto-modelizer-plugin-core';
+import {
+  computed,
   onMounted,
   onUnmounted,
   reactive,
   ref,
-  computed,
 } from 'vue';
-import ComponentDefinitionsDrawer from 'src/components/drawer/ComponentDefinitionsDrawer';
-import ComponentDetailPanel from 'components/drawer/ComponentDetailPanel';
-import {
-  getPluginByName,
-  renderModel,
-  renderConfiguration,
-  addNewComponent,
-  addNewTemplateComponent,
-  initComponents,
-} from 'src/composables/PluginManager';
-import PluginEvent from 'src/composables/events/PluginEvent';
+import { useRoute } from 'vue-router';
 import { Notify } from 'quasar';
 import { useI18n } from 'vue-i18n';
-import { useRoute } from 'vue-router';
-import { getTemplatesByType } from 'src/composables/TemplateManager';
-import ComponentDropOverlay from 'components/drawer/ComponentDropOverlay';
-import { ComponentDrawOption } from 'leto-modelizer-plugin-core';
-
-const route = useRoute();
-const query = computed(() => route.query);
-
-let pluginInitSubscription;
-let pluginDefaultSubscription;
+import PluginEvent from 'src/composables/events/PluginEvent';
 
 const { t } = useI18n();
-const props = defineProps({
-  projectName: {
-    type: String,
-    required: true,
-  },
-});
+const route = useRoute();
+
+const projectName = computed(() => route.params.projectName);
+const query = computed(() => route.query);
 
 const data = reactive({
   plugin: null,
@@ -76,6 +46,9 @@ const templates = ref([]);
 const defaultFolder = ref(process.env.MODELS_DEFAULT_FOLDER !== ''
   ? `${process.env.MODELS_DEFAULT_FOLDER}/`
   : '');
+
+let pluginDefaultSubscription;
+let pluginInitSubscription;
 
 /**
  * On 'Drawer' event type, call renderConfiguration if action is 'move',
@@ -90,57 +63,18 @@ async function onDefaultEvent({ event }) {
   if (event.type === 'Drawer') {
     if (event.action === 'move') {
       await renderConfiguration(
-        props.projectName,
+        projectName.value,
         `${defaultFolder.value}${query.value.path}`,
         data.plugin,
       );
     } else if (renderModelActions.includes(event.action)) {
       await renderModel(
-        props.projectName,
+        projectName.value,
         `${defaultFolder.value}${query.value.path}`,
         data.plugin,
       );
     }
   }
-}
-
-/**
- * Update plugin, draw components and update component templates array.
- * @return {Promise<void>} Promise with nothing on success otherwise an error.
- */
-async function initView() {
-  if (!query.value?.path) {
-    return;
-  }
-
-  const pluginName = query.value.path.split('/')[0];
-  data.plugin = getPluginByName(pluginName);
-
-  if (!data.plugin) {
-    return;
-  }
-
-  await Promise.allSettled([
-    initComponents(
-      route.params.projectName,
-      data.plugin,
-      `${defaultFolder.value}${route.query.path}`,
-    ).then(() => {
-      data.plugin.draw('root');
-    }),
-    getTemplatesByType(
-      'component',
-      data.plugin.data.name,
-    ).then((response) => {
-      templates.value = response;
-    }).catch(() => {
-      Notify.create({
-        type: 'negative',
-        message: t('errors.templates.getData'),
-        html: true,
-      });
-    }),
-  ]);
 }
 
 /**
@@ -204,55 +138,49 @@ async function dropHandler(event) {
 }
 
 onMounted(async () => {
-  pluginInitSubscription = PluginEvent.InitEvent.subscribe(() => {
-    initView();
-  });
   pluginDefaultSubscription = PluginEvent.DefaultEvent.subscribe((event) => {
     onDefaultEvent(event);
   });
-
-  await initView();
+  pluginInitSubscription = PluginEvent.InitEvent.subscribe(() => {
+    data.plugin = getPluginByName(query.value.path.split('/')[0]);
+  });
+  data.plugin = getPluginByName(query.value.path.split('/')[0]);
 });
 
 onUnmounted(() => {
-  pluginInitSubscription.unsubscribe();
   pluginDefaultSubscription.unsubscribe();
+  pluginInitSubscription.unsubscribe();
 });
 </script>
 
 <style scoped>
-  #root {
-    height: calc(100vh - 74px);
-    width: 100%;
-  }
-  .modelizer-draw-view {
-    height: calc(100vh - 64px)
-  }
+#root {
+  height: calc(100vh - 74px);
+  width: 100%;
+}
 </style>
 
 <style lang="scss">
 // Quasar sets overflow to 'hidden' on all svg.
 // In our case, it needs to be set to 'visible' to manage position with % in plugin models.
-  div#root svg {
-    overflow: visible !important;
-    display: unset;
-    height: 100%;
-    width: 100%;
-  }
-
-  #app {
-    font-family: Avenir, Helvetica, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #2c3e50;
-    margin-top: 60px;
-  }
-
-  #viewport {
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    user-select: none;
-  }
+div#root svg {
+  overflow: visible !important;
+  display: unset;
+  height: 100%;
+  width: 100%;
+}
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+  margin-top: 60px;
+}
+#viewport {
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
 </style>