+++ src/router/routes.js
@@ -1,22 +1,56 @@
-import ModelizerRoute from 'src/router/routes/ModelizerRoute';
-import ModelsRoute from 'src/router/routes/ModelsRoute';
+import ModelizerDrawLayout from 'src/layouts/ModelizerDrawLayout.vue';
+import ModelizerTextLayout from 'src/layouts/ModelizerTextLayout.vue';
+import ProjectLayout from 'src/layouts/ProjectLayout.vue';
+import DiagramsLayout from 'src/layouts/DiagramsLayout.vue';
+import AboutLayout from 'src/layouts/AboutLayout.vue';
+import HomeLayout from 'src/layouts/HomeLayout.vue';
+import { getProjectById } from 'src/composables/Project';
 
 const routes = [
   {
     path: '/',
-    component: () => import('layouts/MainLayout.vue'),
+    name: 'Home',
+    component: HomeLayout,
+  },
+  {
+    path: '/projects/:projectName',
+    beforeEnter: (to, _from, next) => {
+      const project = getProjectById(to.params.projectName);
+
+      if (!project) {
+        next({ name: 'Error' });
+      } else {
+        next();
+      }
+    },
     children: [
-      { path: '', name: 'Home', component: () => import('pages/HomePage.vue') },
-      { path: '/about', name: 'About', component: () => import('pages/AboutPage.vue') },
-      ModelizerRoute,
       {
-        path: '/:projectName/diagrams',
+        path: 'modelizer/draw',
+        name: 'Draw',
+        component: ModelizerDrawLayout,
+      },
+      {
+        path: 'modelizer/text',
+        name: 'Text',
+        component: ModelizerTextLayout,
+      },
+      {
+        path: 'models',
+        name: 'Models',
+        component: ProjectLayout,
+      },
+      {
+        path: 'diagrams',
         name: 'Diagrams',
-        component: () => import('pages/DiagramsPage.vue'),
+        component: DiagramsLayout,
       },
     ],
   },
-  ModelsRoute,
+  {
+    path: '/about',
+    name: 'About',
+    component: AboutLayout,
+  },
   // Always leave this as last one,
   // but you can also remove it
   {