+++ tests/unit/pages/DiagramsPage.spec.js
@@ -1,6 +1,7 @@
 import { installQuasarPlugin } from '@quasar/quasar-app-extension-testing-unit-jest';
 import { shallowMount } from '@vue/test-utils';
 import * as PluginManager from 'src/composables/PluginManager';
+import PluginEvent from 'src/composables/events/PluginEvent';
 import DiagramsPage from 'src/pages/DiagramsPage';
 
 installQuasarPlugin();
@@ -19,10 +20,26 @@ jest.mock('src/composables/PluginManager', () => ({
   getPluginByName: jest.fn(() => ({ draw: () => ({}) })),
 }));
 
+jest.mock('src/composables/events/PluginEvent', () => ({
+  InitEvent: {
+    subscribe: jest.fn(),
+  },
+}));
+
 describe('Test page component: DiagramsPage', () => {
   let wrapper;
+  let pluginInitSubscribe;
+  let pluginInitUnsubscribe;
 
   beforeEach(() => {
+    pluginInitSubscribe = jest.fn();
+    pluginInitUnsubscribe = jest.fn();
+
+    PluginEvent.InitEvent.subscribe.mockImplementation(() => {
+      pluginInitSubscribe();
+      return { unsubscribe: pluginInitUnsubscribe };
+    });
+
     wrapper = shallowMount(DiagramsPage);
   });
 
@@ -84,6 +101,20 @@ describe('Test page component: DiagramsPage', () => {
 
       expect(plugins.pluginOne.draw).toHaveBeenCalled();
       expect(plugins.pluginTwo.draw).toHaveBeenCalled();
+    });
+  });
+
+  describe('Test hook function: onMounted', () => {
+    it('should subscribe InitEvent', () => {
+      expect(pluginInitSubscribe).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('Test hook function: onUnmounted', () => {
+    it('should unsubscribe InitEvent', () => {
+      expect(pluginInitUnsubscribe).toHaveBeenCalledTimes(0);
+      wrapper.unmount();
+      expect(pluginInitUnsubscribe).toHaveBeenCalledTimes(1);
     });
   });
 });